Create Opencvbase.py
code to find the time spent in seeing the screen, time spent in seeing far objects and time spent in seeing near objects outside the screen with the respective timestamps, we can use the same approach as before to detect the user's face and facial landmarks, and estimate the distance between the user's face and the camera using the size of the face in the image. Then, we can use the distance information to determine whether the user is looking at the screen or not, and whether they are seeing far or near objects.